requirements.txt
Flask
gunicorn
pandas
numpy
sklearn
datetime
PyYAML
evidently
dill
matplotlib